sam3/requirements.txt
torch
torchvision
numpy
Pillow
sam3 
requests
runpod
psutil  